bug fixes
Not the ideal solution but if it shuts the damn thing up it's worth it. Essentially it would not let me do a ! function call in an if statement so I had to run the function then have it do something cause it won't do nothing in an if statement, then do what I ACTUALLY want in the else statement.